+++ presto-main/src/main/java/io/prestosql/execution/scheduler/NodeScheduler.java
@@ -29,6 +29,10 @@ import io.prestosql.spi.HostAddress;
 
 import javax.inject.Inject;
 
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileReader;
+import java.io.IOException;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.util.ArrayList;
@@ -42,6 +46,8 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.collect.ImmutableList.toImmutableList;
@@ -53,7 +59,7 @@ public class NodeScheduler
 {
     private final NodeSelectorFactory nodeSelectorFactory;
     private static String blackListPath;
-    private static Supplier<List<String>> blackHostList = Suppliers.memoizeWithExpiration(() -> SchedulerUtil.getBlackHosts(blackListPath), 5, TimeUnit.SECONDS);
+    private static Supplier<List<String>> blackHostList = Suppliers.memoizeWithExpiration(() -> getBlackHosts(blackListPath), 5, TimeUnit.SECONDS);
 
     @Inject
     public NodeScheduler(NodeSelectorFactory nodeSelectorFactory)
@@ -227,5 +233,62 @@ public class NodeScheduler
                 .map(remoteTask -> remoteTask.whenSplitQueueHasSpace(spaceThreshold))
                 .collect(toImmutableList());
         return whenAnyCompleteCancelOthers(stateChangeFutures);
+    }
+
+    public static String isReplaceStr(String str)
+    {
+        String r = "";
+        Pattern p = Pattern.compile("\\s*|\t|\r|\n");
+        Matcher m = p.matcher(str);
+        r = m.replaceAll("");
+        return r;
+    }
+
+    public static boolean isIPAddress(String str)
+    {
+        if (str.length() < 7 || str.length() > 15) {
+            return false;
+        }
+        String[] arr = str.split("\\.");
+        if (arr.length != 4) {
+            return false;
+        }
+        for (int i = 0; i < 4; i++) {
+            for (int j = 0; j < arr[i].length(); j++) {
+                char temp = arr[i].charAt(j);
+                if (!(temp > '0' && temp < '9')) {
+                    return false;
+                }
+            }
+        }
+        for (int i = 0; i < 4; i++) {
+            int temp = Integer.parseInt(arr[i]);
+            if (temp < 0 || temp > 255) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    public static List<String> getBlackHosts(String pathname)
+    {
+        //String pathname = blackListPath;
+        List<String> list = new ArrayList<>(8);
+
+        if ((new File(pathname)).exists()) {
+            try (FileReader reader = new FileReader(pathname); BufferedReader br = new BufferedReader(reader)) {
+                String line;
+                while ((line = br.readLine()) != null) {
+                    line = isReplaceStr(line);
+                    if (isIPAddress(line)) {
+                        list.add(line);
+                    }
+                }
+            }
+            catch (IOException e) {
+                throw new RuntimeException("cannot read file: " + pathname, e);
+            }
+        }
+        return list;
     }
 }
